requirements.txt
# Core ML libraries
torch>=2.0.0
transformers>=4.30.0
datasets>=2.12.0
accelerate>=0.20.0
peft>=0.4.0

# Scientific computing
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0

# Data processing
pandas>=2.0.0
tqdm>=4.65.0

# Configuration management
pyyaml>=6.0
omegaconf>=2.3.0

# Evaluation and metrics
nltk>=3.8.0
rouge-score>=0.1.2
bert-score>=0.3.13

# API integrations
openai>=1.0.0

# Experiment tracking
wandb>=0.15.0
tensorboard>=2.13.0

# Utilities
typing-extensions>=4.5.0
dataclasses-json>=0.5.9
click>=8.1.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Development tools
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0

# Optional: For faster tokenization
tokenizers>=0.13.0

# Optional: For model parallelism
deepspeed>=0.9.0